Calculate overall activity summary from epoch data

Calculate overall activity summary from epoch data

activity_summary(
  epochFile,
  nonWearFile = NULL,
  outputFolder = NULL,
  verbose = FALSE,
  timeZone = "Europe/London",
  timeShift = 0,
  startTime = NULL,
  endTime = NULL,
  epochPeriod = 30L,
  stationaryStd = 13L,
  mgCutPointMVPA = 100L,
  mgCutPointVPA = 425L,
  minNonWearDuration = 60L,
  useRecommendedImputation = TRUE,
  activityClassification = TRUE,
  activityModel = c("walmsley-nov20", "doherty-may20", "willetts-may20"),
  model_dir = NULL,
  psd = TRUE,
  fourierFrequency = TRUE,
  fourierWithAcc = TRUE,
  m10l5 = FALSE
)

Arguments

epochFile
Input csv.gz file of processed epoch data

nonWearFile
Output filename for non wear .csv.gz episodes

outputFolder
folder for all of the output files

verbose
print diagnostic messages

timeZone
timezone in country/city format to be used for daylight savings crossover check

timeShift
time shift to be applied, e.g. -15 will shift the device internal time by -15 minutes. Not to be confused with timezone offsets.

startTime
removes data before this time (local) in the final analysis

endTime
removes data after this time (local) in the final analysis

epochPeriod
length in seconds of a single epoch

stationaryStd
stationary mg threshold

mgCutPointMVPA
MVPA threshold for cut point based activity definition

mgCutPointVPA
VPA threshold for cut point based activity

minNonWearDuration
Minimum duration of nonwear events (minutes)

useRecommendedImputation
Highly recommended method to impute missing data using data from other days around the same time

activityClassification
Use pre-trained random forest to predict activity type

activityModel
trained activity model .tar file. If a file is passed, then it will be used, otherwise, it will be passed to download_activity_model

model_dir
directory to download model to, passed to outdir in download_activity_model

psd
Calculate power spectral density for 24 hour circadian period

fourierFrequency
Calculate dominant frequency of sleep for circadian rhythm analysis

fourierWithAcc
Do the Fourier analysis of circadian rhythms (for PSD and Fourier Frequency) with acceleration data instead of sleep signal

m10l5
Calculate relative amplitude of most and least active acceleration periods for circadian rhythm analysis

Value

A list of data

Details

Get overall activity summary from input epoch file. This is achieved by

get interrupt and data error summary vals

check if data occurs at a daylight savings crossover

calculate wear-time statistics, and write nonWear episodes to file

predict activity from features, and add label column

calculate imputation values to replace nan PA metric values

calculate empirical cumulative distribution function of vector magnitudes

derive main movement summaries (overall, weekday/weekend, and hour)

Examples

epoch_file =  system.file("extdata", "ax3_testfile-epoch.csv.gz",
    package = "pycwa")
# \donttest{
  if (pycwa:::have_python_requirements()) {
    sums = pycwa::activity_summary(epoch_file, model_dir = tempdir(), verbose = 2)
  }

#> Warning: pycwa may not work:  packages not found, please try to install using reticulate::py_install(NULL)

# }