Read Axivity CWA file

Read Axivity CWA file

py_convert_cwa(
  file,
  skipCalibration = TRUE,
  outputFolder = NULL,
  verbose = FALSE,
  timeZone = "Europe/London",
  timeShift = 0,
  startTime = NULL,
  endTime = NULL,
  epochPeriod = 30L,
  sampleRate = 100L,
  useFilter = TRUE,
  csvStartTime = NULL,
  csvSampleRate = NULL,
  calOffset = c(0, 0, 0),
  calSlope = c(1, 1, 1),
  calTemp = c(0, 0, 0),
  meanTemp = 20,
  stationaryStd = 13,
  rawDataParser = c("AccelerometerParser", "omconvert"),
  javaHeapSpace = NULL
)

py_read_cwa(file, ...)

Arguments

file
the <.cwa/.cwa.gz> file to process (e.g. sample.cwa.gz). If the file path contains spaces,it must be enclosed in quote marks

skipCalibration
skip calibration?

outputFolder
folder for all of the output files

verbose
enable verbose logging?

timeZone
timezone in country/city format to be used for daylight savings crossover check

timeShift
time shift to be applied, e.g. -15 will shift the device internal time by -15 minutes. Not to be confused with timezone offsets.

startTime
removes data before this time (local) in the final analysis

endTime
removes data after this time (local) in the final analysis

epochPeriod
length in seconds of a single epoch

sampleRate
resample data to n Hz

useFilter
Filter ENMO trunc values?

csvStartTime
start time for csv file when time column is not available (e.g. 2020-01-01T00:01)

csvSampleRate
sample rate for csv file when time column is not available in Hz, or samples/second

calOffset
accelerometer calibration offset

calSlope
accelerometer calibration slope linking offset to temperature

calTemp
mean temperature in degrees Celsius of stationary data for calibration

meanTemp
mean calibration temperature in degrees

stationaryStd
stationary mg threshold

rawDataParser
type of parser to use. If omconvert, this must be installed in your environment path

javaHeapSpace
amount of heap space (in MB) allocated to the java subprocesses, useful for limiting RAM usage

...
not used

Value

A list of files for the output

Examples

file =  system.file("extdata", "ax3_testfile.cwa.gz", package = "pycwa")
print(pycwa:::check_python_requirements())

#> [1] FALSE

if (pycwa:::have_python_requirements()) {
    res = pycwa::py_convert_cwa(file, verbose = 2)
}

#> Warning: pycwa may not work:  packages not found, please try to install using reticulate::py_install(NULL)

# \donttest{
  if (pycwa:::have_python_requirements()) {
    df = pycwa::py_read_cwa(file, startTime = "2019-02-26T10:55:15")
  }

#> Warning: pycwa may not work:  packages not found, please try to install using reticulate::py_install(NULL)

# }